deployment/requirements.txt
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-bootstrap-templates==1.0.6
dash-core-components==2.0.0
dash-gif-component==1.1.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.37.3
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==9.2.0
plotly==5.10.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
scikit-learn==1.0.2
#scipy==1.9.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
tenacity==8.1.0
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
